refactor : 인수테스트 패키지명 변경
